Reactivating leaflet component to create a PR eithout leaflet charts migration.
